Re-organize and cleanup.  Move "no_std" code to "esp" module
