+++ Cargo.toml
@@ -28,10 +28,9 @@ nix = {version = "0.16.1", optional = true}
 
 # Dependencies for Matrix Voice w/ ESP32
 [target.'cfg(target_arch = "xtensa")'.dependencies]
-cty = {version = "0.2"}
-failure = {version = "0.1.7", default-features = false, features = ["derive"]}
+cty = "0.2"
 esp-idf-sys = "0.1.2"
+failure = {version = "0.1.7", default-features = false, features = ["derive"]}
+lazy_static = {version = "1.4", features = ["spin_no_std"]}
 # TODO: setting #[global_allocator] requires #[alloc_error_handler] which requires nightly
 #esp-idf-alloc = "0.1.1"
-#packed_struct = {version = "0.3", default-features = false}
-#packed_struct_codegen = {version = "0.3", default-features = false}
